requirements.txt
# Auto-generated from pyproject.toml for Render deployment
beautifulsoup4>=4.14.2
# resend>=2.0.0  # Commented: Switched to Gmail SMTP. Uncomment to restore Resend.
docling>=2.61.2
feedparser>=6.0.12
flask>=3.0.0
google-genai>=0.2.0
markdown>=3.7.0
markdownify>=0.11.6
openai>=2.7.2
psycopg2-binary>=2.9.11
pydantic>=2.0.0
python-dotenv>=1.2.1
requests>=2.32.5
sqlalchemy>=2.0.44
youtube-transcript-api>=1.2.3
